fixed create note problems

+++ database/migrations/2023_04_25_181702_create_notes_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('notes', function (Blueprint $table) {
             $table->id();
-            $table->integer('valeur');
+            $table->float('valeur');
             $table->foreignId('idexam');
             $table->foreignId('idstagiere');
